refactor : step2의 피드백 바탕 코드 리팩토링

+++ src/main/java/gift/service/ProductService.java
@@ -3,6 +3,8 @@ package gift.service;
 import gift.dto.ProductRequest;
 import gift.dto.ProductResponse;
 import gift.entity.Product;
+import gift.entity.vo.Money;
+import gift.entity.vo.ProductName;
 import gift.exception.ProductNotFoundException;
 import gift.repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -10,8 +12,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
-
 
 @Service
 public class ProductService {
@@ -24,7 +24,7 @@ public class ProductService {
 
     @Transactional
     public ProductResponse addProduct(ProductRequest request) {
-        Product product = new Product(request.name(), request.price(), request.imageUrl());
+        Product product = request.toEntity();
         Product savedProduct = productRepository.save(product);
         return savedProduct.toResponse();
     }
@@ -33,7 +33,12 @@ public class ProductService {
     public ProductResponse updateProduct(Long id, ProductRequest request) {
         Product product = productRepository.findById(id)
                 .orElseThrow(() -> new ProductNotFoundException("해당 ID의 상품을 찾을 수 없습니다: " + id));
-        product.update(request);
+
+        product.update(
+                new ProductName(request.name()),
+                new Money(request.price()),
+                request.imageUrl()
+        );
         return product.toResponse();
     }
 
@@ -53,6 +58,9 @@ public class ProductService {
 
     @Transactional
     public void deleteProduct(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("삭제하려는 상품을 찾을 수 없습니다: " + id);
+        }
         productRepository.deleteById(id);
     }
